app/requirements.txt
Flask
mysql-connector-python
opencv-python-headless
